move ClubListItem raw data to clubListStore
