+++ src/stores/clubListStore/index.js
@@ -1,10 +1,51 @@
 import { observable, action } from 'mobx'
+import React from "react";
 
 const log = console.log.bind(console, '### clubListStore ')
 
+// raw data like this
+const dataList = [{
+    clubName: '慢跑俱乐部',
+    avatar: 'avatar.png',
+    clubID: 1001,
+}, {
+    clubName: '吃饭俱乐部',
+    avatar: 'avatar.png',
+    clubID: 1002,
+},{
+    clubName: '睡觉俱乐部',
+    avatar: 'avatar.png',
+    clubID: 1003,
+},{
+    clubName: '摸鱼俱乐部',
+    avatar: 'avatar.png',
+    clubID: 1004,
+},{
+    clubName: '加班俱乐部',
+    avatar: 'avatar.png',
+    clubID: 1006,
+},{
+    clubName: '加班俱乐部',
+    avatar: 'avatar.png',
+    clubID: 1006,
+},{
+    clubName: '加班俱乐部',
+    avatar: 'avatar.png',
+    clubID: 1006,
+},{
+    clubName: '加班俱乐部',
+    avatar: 'avatar.png',
+    clubID: 1006,
+},];
+
+const arrowRight = <img src="jt_next2.png" alt=""/>
+
 class ClubListStore {
+    @observable dataList
 
     constructor() {
+        this.dataList = dataList
+        this.arrowRight = arrowRight
     }
 
     @action handleClubDetail = () => {
